# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "EstimateResponse": {
        "additionalProperties": true,
        "description": "The routing and cost a create would get, without creating a task.",
        "properties": {
          "budget": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BudgetState"
              },
              {
                "type": "null"
              }
            ]
          },
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/EstimateCandidate"
            },
            "title": "Candidates",
            "type": "array"
          },
          "currency": {
            "const": "USD",
            "default": "USD",
            "title": "Currency",
            "type": "string"
          },
          "estimated_cost": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Cost"
          },
          "modality": {
            "enum": [
              "image",
              "video",
              "music"
            ],
            "title": "Modality",
            "type": "string"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "object": {
            "const": "estimate",
            "default": "estimate",
            "title": "Object",
            "type": "string"
          }
        },
        "required": [
          "modality"
        ],
        "title": "EstimateResponse",
        "type": "object"
      },
...
      "ImageInput": {
        "additionalProperties": false,
        "properties": {
          "type": {
            "const": "image",
            "title": "Type",
            "type": "string"
          },
          "uri": {
            "description": "Absolute media URI. Inline media uses a base64 data URI.",
            "format": "uri",
            "minLength": 1,
            "title": "Uri",
            "type": "string"
          }
        },
        "required": [
          "type",
          "uri"
        ],
        "title": "ImageInput",
        "type": "object"
      },
...
      "ImageRequest": {
        "additionalProperties": false,
        "example": {
          "input": [
            {
              "text": "a cyberpunk cat in the rain",
              "type": "text"
            }
          ],
          "metadata": {
            "requester": "design-tool"
          },
          "model": "gateway-image-pro",
          "parameters": {
            "delivery": "remote",
            "dimensions": {
              "height": 1024,
              "width": 1024
            },
            "quality": "high"
          }
        },
        "properties": {
          "input": {
            "description": "Non-empty ordered image-generation inputs.",
            "items": {
              "discriminator": {
                "mapping": {
                  "image": "#/components/schemas/ImageInput",
                  "text": "#/components/schemas/TextInput"
                },
                "propertyName": "type"
              },
              "oneOf": [
                {
                  "$ref": "#/components/schemas/TextInput"
                },
                {
                  "$ref": "#/components/schemas/ImageInput"
                }
              ]
            },
            "minItems": 1,
            "title": "Input",
            "type": "array"
          },
          "metadata": {
            "additionalProperties": true,
            "description": "Client-owned metadata returned unchanged with the task.",
            "title": "Metadata",
            "type": "object"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Model id returned by GET /v1/models, or omit / set to `auto` to let the gateway auto-route to a backend whose limits fit the request's input (modalities, dimensions, duration, ...).",
            "title": "Model"
          },
          "parameters": {
            "$ref": "#/components/schemas/ImageParameters"
          },
          "routing": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RoutingDirective"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "input"
        ],
        "title": "ImageRequest",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "bearerFormat": "API key",
        "description": "Front-end API key sent as \"Authorization: Bearer <token>\".",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Unified image, video, and music gateway with separate REST APIs.",
    "title": "mm-gateway",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/images/estimate": {
      "post": {
        "operationId": "estimateImage",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EstimateResponse"
                }
              }
            },
            "description": "How auto mode would route the request, and its estimated cost."
          },
          "400": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "error": {
                    "summary": "Invalid request",
                    "value": {
                      "code": "invalid_request_error",
                      "detail": "The routing profile is unavailable.",
                      "instance": "/v1/images/estimate",
                      "request_id": "req_01HZX4J3K7NQ8X2V9Y6R5W4T3P",
                      "status": 400,
                      "title": "Invalid Request Error",
                      "type": "urn:mm-gateway:problem:invalid_request_error"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Invalid request (invalid_request_error / unsupported_feature)."
          },
          "401": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "error": {
                    "summary": "Missing API key",
                    "value": {
                      "code": "unauthorized",
                      "detail": "Missing API key. Send an 'Authorization: Bearer <token>' header.",
                      "instance": "/v1/images/estimate",
                      "request_id": "req_01HZX4J3K7NQ8X2V9Y6R5W4T3P",
                      "status": 401,
                      "title": "Unauthorized",
                      "type": "urn:mm-gateway:problem:unauthorized"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Missing or unknown API key (unauthorized)."
          },
          "403": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "error": {
                    "summary": "Forbidden",
                    "value": {
                      "code": "forbidden",
                      "detail": "The API key is not allowed to perform this request.",
                      "instance": "/v1/images/estimate",
                      "request_id": "req_01HZX4J3K7NQ8X2V9Y6R5W4T3P",
                      "status": 403,
                      "title": "Forbidden",
                      "type": "urn:mm-gateway:problem:forbidden"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Key not allowed to perform the request (forbidden)."
          },
          "404": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "error": {
                    "summary": "Unknown task",
                    "value": {
                      "code": "task_not_found",
                      "detail": "image task img_unknown not found",
                      "instance": "/v1/images/estimate",
                      "request_id": "req_01HZX4J3K7NQ8X2V9Y6R5W4T3P",
                      "status": 404,
                      "title": "Task Not Found",
                      "type": "urn:mm-gateway:problem:task_not_found"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Model or task not found."
          },
          "422": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "error": {
                    "summary": "Validation failed",
                    "value": {
                      "code": "validation_error",
                      "detail": "Request validation failed.",
                      "errors": [
                        {
                          "loc": [
                            "body",
                            "model"
                          ],
                          "msg": "Field required",
                          "type": "missing"
                        }
                      ],
                      "instance": "/v1/images/estimate",
                      "request_id": "req_01HZX4J3K7NQ8X2V9Y6R5W4T3P",
                      "status": 422,
                      "title": "Validation Error",
                      "type": "urn:mm-gateway:problem:validation_error"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Validation Error"
          },
          "502": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "error": {
                    "summary": "Generation service error",
                    "value": {
                      "code": "generation_service_error",
                      "detail": "The generation service returned an error.",
                      "instance": "/v1/images/estimate",
                      "request_id": "req_01HZX4J3K7NQ8X2V9Y6R5W4T3P",
                      "status": 502,
                      "title": "Generation Service Error",
                      "type": "urn:mm-gateway:problem:generation_service_error"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Generation service returned an error."
          },
          "503": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "error": {
                    "summary": "Generation service unavailable",
                    "value": {
                      "code": "generation_service_unavailable",
                      "detail": "No generation service is available for this request.",
                      "instance": "/v1/images/estimate",
                      "request_id": "req_01HZX4J3K7NQ8X2V9Y6R5W4T3P",
                      "status": 503,
                      "title": "Generation Service Unavailable",
                      "type": "urn:mm-gateway:problem:generation_service_unavailable"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "No usable generation service is configured."
          },
          "504": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "error": {
                    "summary": "Generation service timeout",
                    "value": {
                      "code": "generation_service_timeout",
                      "detail": "The generation service timed out.",
                      "instance": "/v1/images/estimate",
                      "request_id": "req_01HZX4J3K7NQ8X2V9Y6R5W4T3P",
                      "status": 504,
                      "title": "Generation Service Timeout",
                      "type": "urn:mm-gateway:problem:generation_service_timeout"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Generation service timed out."
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Estimate an image request",
        "tags": [
          "images"
        ]
      }
    },
...
    "/v1/music/estimate": {
      "post": {
        "operationId": "estimateMusic",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MusicRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EstimateResponse"
                }
              }
            },
            "description": "How auto mode would route the request, and its estimated cost."
          },
          "400": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "error": {
                    "summary": "Invalid request",
                    "value": {
                      "code": "invalid_request_error",
                      "detail": "The routing profile is unavailable.",
                      "instance": "/v1/music/estimate",
                      "request_id": "req_01HZX4J3K7NQ8X2V9Y6R5W4T3P",
                      "status": 400,
                      "title": "Invalid Request Error",
                      "type": "urn:mm-gateway:problem:invalid_request_error"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Invalid request (invalid_request_error / unsupported_feature)."
          },
          "401": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "error": {
                    "summary": "Missing API key",
                    "value": {
                      "code": "unauthorized",
                      "detail": "Missing API key. Send an 'Authorization: Bearer <token>' header.",
                      "instance": "/v1/music/estimate",
                      "request_id": "req_01HZX4J3K7NQ8X2V9Y6R5W4T3P",
                      "status": 401,
                      "title": "Unauthorized",
                      "type": "urn:mm-gateway:problem:unauthorized"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Missing or unknown API key (unauthorized)."
          },
          "403": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "error": {
                    "summary": "Forbidden",
                    "value": {
                      "code": "forbidden",
                      "detail": "The API key is not allowed to perform this request.",
                      "instance": "/v1/music/estimate",
                      "request_id": "req_01HZX4J3K7NQ8X2V9Y6R5W4T3P",
                      "status": 403,
                      "title": "Forbidden",
                      "type": "urn:mm-gateway:problem:forbidden"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Key not allowed to perform the request (forbidden)."
          },
          "404": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "error": {
                    "summary": "Unknown task",
                    "value": {
                      "code": "task_not_found",
                      "detail": "image task img_unknown not found",
                      "instance": "/v1/music/estimate",
                      "request_id": "req_01HZX4J3K7NQ8X2V9Y6R5W4T3P",
                      "status": 404,
                      "title": "Task Not Found",
                      "type": "urn:mm-gateway:problem:task_not_found"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Model or task not found."
          },
          "422": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "error": {
                    "summary": "Validation failed",
                    "value": {
                      "code": "validation_error",
                      "detail": "Request validation failed.",
                      "errors": [
                        {
                          "loc": [
                            "body",
                            "model"
                          ],
                          "msg": "Field required",
                          "type": "missing"
                        }
                      ],
                      "instance": "/v1/music/estimate",
                      "request_id": "req_01HZX4J3K7NQ8X2V9Y6R5W4T3P",
                      "status": 422,
                      "title": "Validation Error",
                      "type": "urn:mm-gateway:problem:validation_error"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Validation Error"
          },
          "502": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "error": {
                    "summary": "Generation service error",
                    "value": {
                      "code": "generation_service_error",
                      "detail": "The generation service returned an error.",
                      "instance": "/v1/music/estimate",
                      "request_id": "req_01HZX4J3K7NQ8X2V9Y6R5W4T3P",
                      "status": 502,
                      "title": "Generation Service Error",
                      "type": "urn:mm-gateway:problem:generation_service_error"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Generation service returned an error."
          },
          "503": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "error": {
                    "summary": "Generation service unavailable",
                    "value": {
                      "code": "generation_service_unavailable",
                      "detail": "No generation service is available for this request.",
                      "instance": "/v1/music/estimate",
                      "request_id": "req_01HZX4J3K7NQ8X2V9Y6R5W4T3P",
                      "status": 503,
                      "title": "Generation Service Unavailable",
                      "type": "urn:mm-gateway:problem:generation_service_unavailable"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "No usable generation service is configured."
          },
          "504": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "error": {
                    "summary": "Generation service timeout",
                    "value": {
                      "code": "generation_service_timeout",
                      "detail": "The generation service timed out.",
                      "instance": "/v1/music/estimate",
                      "request_id": "req_01HZX4J3K7NQ8X2V9Y6R5W4T3P",
                      "status": 504,
                      "title": "Generation Service Timeout",
                      "type": "urn:mm-gateway:problem:generation_service_timeout"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Generation service timed out."
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Estimate a music request",
        "tags": [
          "music"
        ]
      }
    },
...
    "/v1/usage": {
      "get": {
        "operationId": "getUsage",
        "parameters": [
          {
            "description": "Report only this budget scope.",
            "in": "query",
            "name": "scope",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "minLength": 1,
                  "pattern": "^[A-Za-z0-9._:/-]+$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Report only this budget scope.",
              "title": "Scope"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsageResponse"
                }
              }
            },
            "description": "Spend, reservations and budgets for the current period."
          },
          "400": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "error": {
                    "summary": "Invalid request",
                    "value": {
                      "code": "invalid_request_error",
                      "detail": "The routing profile is unavailable.",
                      "instance": "/v1/usage",
                      "request_id": "req_01HZX4J3K7NQ8X2V9Y6R5W4T3P",
                      "status": 400,
                      "title": "Invalid Request Error",
                      "type": "urn:mm-gateway:problem:invalid_request_error"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Invalid request (invalid_request_error / unsupported_feature)."
          },
          "401": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "error": {
                    "summary": "Missing API key",
                    "value": {
                      "code": "unauthorized",
                      "detail": "Missing API key. Send an 'Authorization: Bearer <token>' header.",
                      "instance": "/v1/usage",
                      "request_id": "req_01HZX4J3K7NQ8X2V9Y6R5W4T3P",
                      "status": 401,
                      "title": "Unauthorized",
                      "type": "urn:mm-gateway:problem:unauthorized"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Missing or unknown API key (unauthorized)."
          },
          "403": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "error": {
                    "summary": "Forbidden",
                    "value": {
                      "code": "forbidden",
                      "detail": "The API key is not allowed to perform this request.",
                      "instance": "/v1/usage",
                      "request_id": "req_01HZX4J3K7NQ8X2V9Y6R5W4T3P",
                      "status": 403,
                      "title": "Forbidden",
                      "type": "urn:mm-gateway:problem:forbidden"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Key not allowed to perform the request (forbidden)."
          },
          "404": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "error": {
                    "summary": "Unknown task",
                    "value": {
                      "code": "task_not_found",
                      "detail": "image task img_unknown not found",
                      "instance": "/v1/usage",
                      "request_id": "req_01HZX4J3K7NQ8X2V9Y6R5W4T3P",
                      "status": 404,
                      "title": "Task Not Found",
                      "type": "urn:mm-gateway:problem:task_not_found"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Model or task not found."
          },
          "422": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "error": {
                    "summary": "Validation failed",
                    "value": {
                      "code": "validation_error",
                      "detail": "Request validation failed.",
                      "errors": [
                        {
                          "loc": [
                            "body",
                            "model"
                          ],
                          "msg": "Field required",
                          "type": "missing"
                        }
                      ],
                      "instance": "/v1/usage",
                      "request_id": "req_01HZX4J3K7NQ8X2V9Y6R5W4T3P",
                      "status": 422,
                      "title": "Validation Error",
                      "type": "urn:mm-gateway:problem:validation_error"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Validation Error"
          },
          "502": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "error": {
                    "summary": "Generation service error",
                    "value": {
                      "code": "generation_service_error",
                      "detail": "The generation service returned an error.",
                      "instance": "/v1/usage",
                      "request_id": "req_01HZX4J3K7NQ8X2V9Y6R5W4T3P",
                      "status": 502,
                      "title": "Generation Service Error",
                      "type": "urn:mm-gateway:problem:generation_service_error"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Generation service returned an error."
          },
          "503": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "error": {
                    "summary": "Generation service unavailable",
                    "value": {
                      "code": "generation_service_unavailable",
                      "detail": "No generation service is available for this request.",
                      "instance": "/v1/usage",
                      "request_id": "req_01HZX4J3K7NQ8X2V9Y6R5W4T3P",
                      "status": 503,
                      "title": "Generation Service Unavailable",
                      "type": "urn:mm-gateway:problem:generation_service_unavailable"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "No usable generation service is configured."
          },
          "504": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "error": {
                    "summary": "Generation service timeout",
                    "value": {
                      "code": "generation_service_timeout",
                      "detail": "The generation service timed out.",
                      "instance": "/v1/usage",
                      "request_id": "req_01HZX4J3K7NQ8X2V9Y6R5W4T3P",
                      "status": 504,
                      "title": "Generation Service Timeout",
                      "type": "urn:mm-gateway:problem:generation_service_timeout"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Generation service timed out."
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Spend and budgets of the authenticated key",
        "tags": [
          "usage"
        ]
      }
    },
...
  }
}